fix #400 (#413)
* WIP

* WIP2

* Basic support for gzip and bzip2

* Fix the inconsistency between Linux and Windows

* Test also import

* Remove `which` for `.parse_rutils`

* Remove the useless checks

* Doc update

* Remove the `which` warning for multiple file archive

* Fix #396

* Update doc and remove commented out lines [no ci]

+++ R/compression.R
@@ -5,26 +5,42 @@ find_compress <- function(f) {
     if (grepl("\\.tar\\.gz$", f)) {
         return(list(file = sub("\\.tar\\.gz$", "", f), compress = "tar.gz"))
     }
+    if (grepl(".tgz$", f)) {
+        return(list(file = sub("\\.tgz$", "", f), compress = "tar.gz"))
+    }
+    if (grepl("\\.tar\\.bz2$", f)) {
+        return(list(file = sub("\\.tar\\.bz2$", "", f), compress = "tar.bz2"))
+    }
+    if (grepl("\\.tbz2$", f)) {
+        return(list(file = sub("\\.tbz2$", "", f), compress = "tar.bz2"))
+    }
     if (grepl("\\.tar$", f)) {
         return(list(file = sub("\\.tar$", "", f), compress = "tar"))
+    }
+    if (grepl("\\.gzip$", f)) {
+        ## weird
+        return(list(file = sub("\\.gzip$", "", f), compress = "gzip"))
+    }
+    if (grepl("\\.gz$", f)) {
+        return(list(file = sub("\\.gz$", "", f), compress = "gzip"))
+    }
+    if (grepl("\\.bz2$", f)) {
+        return(list(file = sub("\\.bz2$", "", f), compress = "bzip2"))
+    }
+    if (grepl("\\.bzip2$", f)) {
+        ## weird
+        return(list(file = sub("\\.bzip2$", "", f), compress = "bzip2"))
     }
     return(list(file = f, compress = NA_character_))
 }
 
-## KEEPING OLD CODE FOR LATER REIMPLEMENTATION for gzip and bzip2 #400
-##compress_out <- function(cfile, filename, type = c("zip", "tar", "gzip", "bzip2", "xz")) {
-compress_out <- function(cfile, filename, type = c("zip", "tar", "tar.gz")) {
+compress_out <- function(cfile, filename, type = c("zip", "tar", "tar.gz", "tar.bz2", "gzip", "bzip2")) {
     type <- ext <- match.arg(type)
-    ## if (ext %in% c("gzip", "bzip2", "xz")) {
-    ##     ext <- paste0("tar")
-    ## }
-    if (missing(cfile)) {
-        cfile <- paste0(filename, ".", ext)
-        cfile2 <- paste0(basename(filename), ".", ext)
-    } else {
-        cfile2 <- basename(cfile)
-    }
+    cfile2 <- basename(cfile)
     filename <- normalizePath(filename)
+    if (type %in% c("gzip", "bzip2")) {
+        return(.compress_rutils(filename, cfile, ext = ext))
+    }
     tmp <- tempfile()
     dir.create(tmp)
     on.exit(unlink(tmp, recursive = TRUE), add = TRUE)
@@ -41,6 +57,9 @@ compress_out <- function(cfile, filename, type = c("zip", "tar", "tar.gz")) {
     if (type == "tar.gz") {
         o <- utils::tar(cfile2, files = basename(filename), compression = "gzip")
     }
+    if (type == "tar.bz2") {
+        o <- utils::tar(cfile2, files = basename(filename), compression = "bzip2")
+    }
     setwd(wd)
     if (o != 0) {
         stop(sprintf("File compression failed for %s!", cfile))
@@ -51,27 +70,21 @@ compress_out <- function(cfile, filename, type = c("zip", "tar", "tar.gz")) {
 }
 
 parse_archive <- function(file, which, file_type, ...) {
-    if (!file_type %in% c("zip", "tar", "tar.gz")) {
-        stop("Unsupported file_type. Use 'zip', 'tar', or 'tar.gz'.")
+    if (file_type %in% c("gzip", "bzip2")) {
+        ## it doesn't have the same interface as unzip
+        return(.parse_rutils(filename = file, file_type = file_type))
     }
     if (file_type == "zip") {
         file_list <- utils::unzip(file, list = TRUE)$Name
         extract_func <- utils::unzip
     }
-    if (file_type %in% c("tar", "tar.gz")) {
+    if (file_type %in% c("tar", "tar.gz", "tar.bz2")) {
         file_list <- utils::untar(file, list = TRUE)
         extract_func <- utils::untar
     }
 
     d <- tempfile()
     dir.create(d)
-
-    if (missing(which)) {
-        if (length(file_list) > 1) {
-            warning(sprintf("%s archive contains multiple files. Attempting first file.", file_type))
-        }
-        which <- 1
-    }
 
     if (is.numeric(which)) {
         extract_func(file, files = file_list[which], exdir = d)
@@ -82,4 +95,26 @@ parse_archive <- function(file, which, file_type, ...) {
     }
     extract_func(file, files = file_list[grep(which2, file_list)[1]], exdir = d)
     return(file.path(d, which))
+}
+
+.compress_rutils <- function(filename, cfile, ext, remove = TRUE, FUN = gzfile) {
+    ## Caution: Please note that remove = TRUE by default, it will delete `filename`!
+    if (ext == "bzip2") {
+        FUN <- bzfile
+    }
+    tmp_cfile <- R.utils::compressFile(filename = filename, destname = tempfile(), ext = ext, FUN = FUN, overwrite = TRUE, remove = remove)
+    file.copy(from = tmp_cfile, to = cfile, overwrite = TRUE)
+    unlink(tmp_cfile)
+    return(cfile)
+}
+
+.parse_rutils <- function(filename, file_type) {
+    if (file_type == "gzip") {
+        decompression_fun <- gzfile
+    }
+    if (file_type == "bzip2") {
+        decompression_fun <- bzfile
+    }
+    destname <- tempfile()
+    R.utils::decompressFile(filename = filename, destname = destname, temporary = TRUE, remove = FALSE, overwrite = TRUE, FUN = decompression_fun, ext = file_type)
 }
